Module p2p::behaviour::firewall

Re-exports

pub use permissions::*;

Modules

permissions

Structs

FirewallConfiguration
Configuration for the firewall of the NetBehaviour. This config specifies what inbound and requests from/ to which peer are allowed. If there are neither default rules, nor a peer specific rule for a request from/ to a peer, a FirewallRequest::PeerSpecificRule will be sent through the firewall-channel that is passed to StrongholdP2p.

FirewallRules
Rule configuration for inbound and outbound requests.

Enums

FirewallRequest
Requests for approval and rules that are not covered by the current FirewallConfiguration.

Rule
Rules for requests in a specific RequestDirection.

RuleDirection
The direction for which a rule is applicable.