Struct p2p::behaviour::firewall::FirewallConfiguration

pub struct FirewallConfiguration<TRq: Clone> {
    default: FirewallRules<TRq>,
    peer_rules: HashMap<PeerId, FirewallRules<TRq>>,
}

Configuration for the firewall of the NetBehaviour. This config specifies what inbound and requests from/ to which peer are allowed. If there are neither default rules, nor a peer specific rule for a request from/ to a peer, a FirewallRequest::PeerSpecificRule will be sent through the firewall-channel that is passed to StrongholdP2p.

Fields
default: FirewallRules<TRq>

Default rules that are used if there are no peer-specific ones for a peer.

peer_rules: HashMap<PeerId, FirewallRules<TRq>>

Peer specific rules.

Implementations

impl<TRq: Clone> FirewallConfiguration<TRq>

pub fn new(
default_in: Option<Rule<TRq>>,
default_out: Option<Rule<TRq>>
) -> Self

Create a new instance with the given default rules. If no rules are set, a a FirewallRequest::PeerSpecificRule will be sent through the firewall-channel on inbound and outbound requests.

pub fn allow_all() -> Self

Create a new instance with default configuration allowing all requests.

pub fn reject_all() -> Self

Create a new instance with default configuration rejecting all requests.

pub fn get_default_rules(&self) -> &FirewallRules<TRq>

Get default firewall rules that are used if there are no peer-specific ones for a direction.

pub fn set_default(
&mut self,
default: Option<Rule<TRq>>,
direction: RuleDirection
)

Set the default rules for one or both directions. In case of None, the rule(s) are removed

pub fn get_rules(&self, peer: &PeerId) -> Option<&FirewallRules<TRq>>

Get the peer specific rules.

pub fn get_effective_rules(&self, peer: &PeerId) -> FirewallRules<TRq>

Get effective rules for a peer i.g. peer-specific rules or else the default rules for each direction.

pub fn set_rule(
&mut self,
peer: PeerId,
rule: Option<Rule<TRq>>,
direction: RuleDirection
)

Set one or both rules for a specific peer. In case of None, the rule(s) are removed.