Struct p2p::behaviour::NetBehaviour

pub struct NetBehaviour<Rq, Rs, TRq> where
    Rq: RqRsMessage + Borrow<TRq>,
    Rs: RqRsMessage,
    TRq: Clone + Send + 'static,  {

    mdns: Mdns,
    relay: Relay,
    supported_protocols: SmallVec<[MessageProtocol; 2]>,
    request_timeout: Duration,
    connection_timeout: Duration,
    next_request_id: RequestId,
    next_inbound_id: Arc<AtomicU64>,
    request_manager: RequestManager<Rq, Rs, TRq>,
    addresses: AddressInfo,
    firewall: FirewallConfiguration<TRq>,
    permission_req_channel: Sender<FirewallRequest<TRq>>,
    pending_rule_rqs: FuturesUnordered<BoxFuture<'static, (PeerId, Option<FirewallRules<TRq>>)>>,
    pending_approval_rqs: FuturesUnordered<BoxFuture<'static, (RequestId, bool)>>,

}

Protocol for customization for the [Swarm][libp2p::Swarm].

The protocol is based on the [RequestResponse][https://docs.rs/libp2p-request-response] protocol from libp2p and integrates the libp2p [Relay][libp2p::relay::Relay] and [Mdns][libp2p::mdns::Mdns] protocols.

This allows sending request messages to remote peers, handling of inbound requests and failures, and additionally the configuration of a firewall to set permissions individually for different peers and request types.

Fields
mdns: Mdnsrelay: Relaysupported_protocols: SmallVec<[MessageProtocol; 2]>request_timeout: Durationconnection_timeout: Durationnext_request_id: RequestIdnext_inbound_id: Arc<AtomicU64>request_manager: RequestManager<Rq, Rs, TRq>addresses: AddressInfofirewall: FirewallConfiguration<TRq>permission_req_channel: Sender<FirewallRequest<TRq>>pending_rule_rqs: FuturesUnordered<BoxFuture<'static, (PeerId, Option<FirewallRules<TRq>>)>>pending_approval_rqs: FuturesUnordered<BoxFuture<'static, (RequestId, bool)>>
Implementations

impl<Rq, Rs, TRq> NetBehaviour<Rq, Rs, TRq> where
Rq: RqRsMessage + Borrow<TRq>,
Rs: RqRsMessage,
TRq: Clone + Send + 'static,

pub fn new(
config: NetBehaviourConfig<TRq>,
mdns: Mdns,
relay: Relay,
permission_req_channel: Sender<FirewallRequest<TRq>>
) -> Self

Create a new instance of a NetBehaviour to customize the [Swarm][libp2p::Swarm].

pub fn send_request(&mut self, peer: PeerId, request: Rq) -> RequestId

Send a new request to a remote peer.

pub fn get_firewall_default(&self) -> &FirewallRules<TRq>

Get the current default rules for the firewall. The default rules are used for peers that do not have any explicit rules.

pub fn set_firewall_default(
&mut self,
direction: RuleDirection,
default: Rule<TRq>
)

Set the default configuration for the firewall. The default rules are used for peers that do not have any explicit rules.

pub fn remove_firewall_default(&mut self, direction: RuleDirection)

Remove a default firewall rule. If there is no default rule and no peer-specific rule, a FirewallRequest::PeerSpecificRule request will be sent through the firewall channel

pub fn get_peer_rules(&self, peer: &PeerId) -> Option<&FirewallRules<TRq>>

Get the explicit rules for a peer, if there are any.

pub fn set_peer_rule(
&mut self,
peer: PeerId,
direction: RuleDirection,
rule: Rule<TRq>
)

Set a peer specific rule to overwrite the default behaviour for that peer.

pub fn remove_peer_rule(&mut self, peer: PeerId, direction: RuleDirection)

Remove a peer specific rule, which will result in using the firewall default rules.

pub fn add_address(&mut self, peer: PeerId, address: Multiaddr)

Add an address for the remote peer.

pub fn remove_address(&mut self, peer: &PeerId, address: &Multiaddr)

Remove an address from the known addresses of a remote peer.

pub fn get_established_connections(
&self
) -> Vec<(PeerId, EstablishedConnections)>

fn next_request_id(&mut self) -> RequestId

RequestId for the next outbound request.

fn check_approval_status(
&mut self,
peer: PeerId,
request_id: RequestId,
request: &TRq,
direction: RequestDirection
) -> ApprovalStatus

fn handle_handler_event(
&mut self,
peer: PeerId,
connection: ConnectionId,
event: HandlerOutEvent<Rq, Rs>
)

fn query_peer_rule(&mut self, peer: PeerId)

fn query_request_approval(
&mut self,
peer: PeerId,
request_id: RequestId,
rq: TRq,
direction: RequestDirection
)

async fn send_firewall(
channel: Sender<FirewallRequest<TRq>>,
request: FirewallRequest<TRq>
) -> Result<(), SendError>

fn handle_updated_peer_rule(&mut self, peer: PeerId, direction: RuleDirection)

impl<Rq, Rs, TRq> NetBehaviour<Rq, Rs, TRq> where
Rq: RqRsMessage + Borrow<TRq>,
Rs: RqRsMessage,
TRq: Clone + Send + 'static,

pub fn add_dialing_relay(
&mut self,
peer: PeerId,
address: Option<Multiaddr>
) -> Option<Multiaddr>

Add a relay to the list of relays that may be tried to use if a remote peer can not be reached directly.

pub fn remove_dialing_relay(&mut self, peer: &PeerId)

Remove a relay from the list of dialing relays.

pub fn set_relay_fallback(&mut self, peer: PeerId, use_relay_fallback: bool)

Configure whether it should be attempted to reach the remote via known relays, if it can not be reached via known addresses.

pub fn use_specific_relay(
&mut self,
target: PeerId,
relay: PeerId,
is_exclusive: bool
) -> Option<Multiaddr>

Dial the target via the specified relay. The is_exclusive paramter specifies whether other known relays should be used if using the set relay is not successful.

Returns the relayed address of the local peer (`//p2p-circuit/), if an address for the relay is known.