Enum p2p::NetworkEvent

pub enum NetworkEvent {
    InboundFailure {
        request_id: RequestId,
        peer: PeerId,
        failure: InboundFailure,
    },
    ConnectionEstablished {
        peer: PeerId,
        endpoint: ConnectedPoint,
        num_established: NonZeroU32,
    },
    ConnectionClosed {
        peer: PeerId,
        endpoint: ConnectedPoint,
        num_established: u32,
        cause: Option<Error>,
    },
    IncomingConnectionError {
        local_addr: Multiaddr,
        send_back_addr: Multiaddr,
        error: ConnectionErr,
    },
    NewListenAddr(Multiaddr),
    ExpiredListenAddr(Multiaddr),
    ListenerClosed {
        addresses: Vec<Multiaddr>,
        cause: Option<Error>,
    },
    ListenerError {
        error: Error,
    },
}

Events happening in the Network. Includes events about connection and listener status as well as potential failures when receiving request-response messages.

Variants

InboundFailure

A failure occurred in the context of receiving an inbound request and sending a response.

Fields of InboundFailure

request_id: RequestIdpeer: PeerIdfailure: InboundFailure

ConnectionEstablished

A connection to the given peer has been opened.

Fields of ConnectionEstablished

peer: PeerId

Identity of the peer that connected.

endpoint: ConnectedPoint

Endpoint of the connection that has been opened.

num_established: NonZeroU32

Number of established connections to this peer, including the one that has just been opened.

ConnectionClosed

A connection with the given peer has been closed, possibly as a result of an error.

Fields of ConnectionClosed

peer: PeerId

Identity of the peer that disconnected.

endpoint: ConnectedPoint

Endpoint of the connection that has been closed.

num_established: u32

Number of other remaining connections to this same peer.

cause: Option<Error>

Potential Error that resulted in the disconnection.

IncomingConnectionError

An error happened on a connection during its initial handshake.

This can include, for example, an error during the handshake of the encryption layer, or the connection unexpectedly closed.

Fields of IncomingConnectionError

local_addr: Multiaddr

Local connection address. This address has been earlier reported with a NewListenAddr event.

send_back_addr: Multiaddr

Address used to send back data to the remote.

error: ConnectionErr

The error that happened.

NewListenAddr(Multiaddr)

One of the listeners has reported a new local listening address.

ExpiredListenAddr(Multiaddr)

One of the listeners has reported the expiration of a listening address.

ListenerClosed

One of the listeners gracefully closed.

Fields of ListenerClosed

addresses: Vec<Multiaddr>

The addresses that the listener was listening on. These addresses are now considered expired, similar to if a ExpiredListenAddr event has been generated for each of them.

cause: Option<Error>

Potential Error in the stream that cause the listener to close.

ListenerError

One of the listeners reported a non-fatal error.

Fields of ListenerError

error: Error

The listener error.