Module p2p::interface::msg_channel

Structs

EventChannel
Wrapper of a [mpsc::channel][futures::channel::mpsc::channel] for sending events. The EventChannel cen be configure its behaviour e capacity of the channel is reached.

Enums

ChannelSinkConfig
Configure how the network should behave in the case that the capacity of the EventChannel is reached, i.g. if the channel is full. This is relevant in cases when the frequency of messages is larger than the frequency in which the [mpsc::Receiver][futures::channel::mpsc::Receiver] side reads from the stream.