Struct p2p::interface::StrongholdP2p

pub struct StrongholdP2p<Rq, Rs, TRq = Rq> where
    Rq: RqRsMessage + Borrow<TRq>,
    Rs: RqRsMessage,
    TRq: Clone + Send + 'static,  {
    local_peer_id: PeerId,
    command_tx: Sender<SwarmOperation<Rq, Rs, TRq>>,
}

Interface for the stronghold-p2p library to create a swarm, handle events and perform operations.

All Swarm interaction takes place in a separate task. StrongholdP2p is essentially a wrapper for the Sender side of a mpsc channel, which is used to initiate operations on the swarm.

Refer to StrongholdP2pBuilder for more information on the default configuration.

Fields
local_peer_id: PeerIdcommand_tx: Sender<SwarmOperation<Rq, Rs, TRq>>
Implementations

impl<Rq, Rs, TRq> StrongholdP2p<Rq, Rs, TRq> where
Rq: RqRsMessage + Borrow<TRq>,
Rs: RqRsMessage,
TRq: Clone + Send + 'static,

pub async fn new(
firewall_channel: Sender<FirewallRequest<TRq>>,
requests_channel: EventChannel<ReceiveRequest<Rq, Rs>>,
events_channel: Option<EventChannel<NetworkEvent>>
) -> Result<Self, Error>

Create a new StrongholdP2p instance with the default configuration. Refer to StrongholdP2pBuilder::new and StrongholdP2pBuilder::build for detailed information.

pub fn get_peer_id(&self) -> PeerId

Get the PeerId of the local peer.

pub async fn send_request(
&mut self,
peer: PeerId,
request: Rq
) -> Result<Rs, OutboundFailure>

Send a new request to a remote peer.

This will attempt to establish a connection to the remote via one of the known addresses, if there is no active connection.

pub async fn start_listening(
&mut self,
address: Multiaddr
) -> Result<Multiaddr, ListenErr>

Start listening on the network on the given address. In case of a tcp-transport, the address /ip4/0.0.0.0/tcp/0 can be set if an OS-assigned address should be used.

Note: Depending on the used transport, this may produce multiple listening addresses. This method only returns the first reported listening address for the new listener. All active listening addresses for each listener can be obtained from StrongholdP2p::get_listeners

pub async fn start_relayed_listening(
&mut self,
relay: PeerId,
relay_addr: Option<Multiaddr>
) -> Result<Multiaddr, ListenRelayErr>

Start listening via a relay peer on an address following the scheme <relay-addr>/<relay-id>/p2p-circuit/<local-id>. This will establish a keep-alive connection to the relay, the relay will forward all requests to the local peer.

pub async fn get_listeners(&mut self) -> Vec<Listener>

pub async fn stop_listening(&mut self)

Stop listening on all listeners.

pub async fn stop_listening_addr(&mut self, address: Multiaddr)

Stop listening on the listener associated with the given address.

pub async fn stop_listening_relay(&mut self, relay: PeerId)

Stop listening via the given relay.

pub async fn connect_peer(&mut self, peer: PeerId) -> Result<Multiaddr, DialErr>

Establish a new new connection to the remote peer. This will try each known address until either a connection was successful, or all failed.

pub async fn set_firewall_default(
&mut self,
direction: RuleDirection,
default: Rule<TRq>
)

Set the default configuration for the firewall. The default rules are used for peers that do not have any explicit rules.

pub async fn get_firewall_default(&mut self) -> FirewallRules<TRq>

Get the current default rules for the firewall. The default rules are used for peers that do not have any explicit rules.

pub async fn remove_firewall_default(&mut self, direction: RuleDirection)

Remove a default firewall rule. If there is no default rule and no peer-specific rule, a FirewallRequest::PeerSpecificRule request will be sent through the firewall channel

pub async fn get_peer_rules(&mut self, peer: PeerId) -> FirewallRules<TRq>

Get the explicit rules for a peer, if there are any.

pub async fn set_peer_rule(
&mut self,
peer: PeerId,
direction: RuleDirection,
rule: Rule<TRq>
)

Set a peer specific rule to overwrite the default behaviour for that peer.

pub async fn remove_peer_rule(&mut self, peer: PeerId, direction: RuleDirection)

Remove a peer specific rule, which will result in using the firewall default rules.

pub async fn get_addrs(&mut self, peer: PeerId) -> Vec<Multiaddr>

Get the known addresses for a remote peer.

pub async fn add_address(&mut self, peer: PeerId, address: Multiaddr)

Add an address for the remote peer.

pub async fn remove_address(&mut self, peer: PeerId, address: Multiaddr)

Remove an address from the known addresses of a remote peer.

pub async fn add_dialing_relay(
&mut self,
peer: PeerId,
address: Option<Multiaddr>
) -> Option<Multiaddr>

Add a relay to the list of relays that may be tried to use if a remote peer can not be reached directly.

pub async fn remove_dialing_relay(&mut self, peer: PeerId)

Remove a relay from the list of dialing relays.

pub async fn set_relay_fallback(
&mut self,
peer: PeerId,
use_relay_fallback: bool
)

Configure whether it should be attempted to reach the remote via known relays, if it can not be reached via known addresses.

pub async fn use_specific_relay(
&mut self,
target: PeerId,
relay: PeerId,
is_exclusive: bool
) -> Option<Multiaddr>

Dial the target via the specified relay. The is_exclusive parameter specifies whether other known relays should be used if using the set relay is not successful.

Returns the relayed address of the local peer (`//p2p-circuit/), if an address for the relay is known.

pub async fn ban_peer(&mut self, peer: PeerId)

Bans a peer by its peer ID.

Any incoming connection and any dialing attempt will immediately be rejected. This function has no effect if the peer is already banned.

pub async fn unban_peer(&mut self, peer: PeerId)

Unbans a peer.

pub async fn is_connected(&mut self, peer: PeerId) -> bool

Check whether the Network has an established connection to a peer.

pub async fn get_connections(&mut self) -> Vec<(PeerId, EstablishedConnections)>

async fn send_command(&mut self, command: SwarmOperation<Rq, Rs, TRq>)