Struct p2p::interface::StrongholdP2pBuilder

pub struct StrongholdP2pBuilder<Rq, Rs, TRq = Rq> where
    Rq: RqRsMessage + Borrow<TRq>,
    Rs: RqRsMessage,
    TRq: Clone + Send + 'static,  {
    firewall_channel: Sender<FirewallRequest<TRq>>,
    requests_channel: EventChannel<ReceiveRequest<Rq, Rs>>,
    events_channel: Option<EventChannel<NetworkEvent>>,
    ident: Option<(AuthenticKeypair<X25519Spec>, PeerId)>,
    behaviour_config: NetBehaviourConfig<TRq>,
    connections_limit: Option<ConnectionLimits>,
}

Builder for new StrongholdP2p.

Default behaviour:

No firewall rules are set. In case of inbound / outbound requests, a FirewallRequest::PeerSpecificRule request is sent through the firewall_channel to specify the rules for this peer.

A new keypair is created and used, from which the PeerId of the local peer is derived.

No limit for simultaneous connections.

Request-timeout and Connection-timeout are 10s.

StrongholdP2p is build either via StrongholdP2pBuilder::build (requires feature tcp-transport) with a pre-configured transport, or StrongholdP2pBuilder::build_with_transport with a custom transport.

When building a new StrongholdP2p a new [Swarm][libp2p::Swarm] is created and continuously polled for events. Inbound requests are forwarded through a mpsc::channel<ReceiveRequest<Rq, Rs>>. Optionally all events regarding connections and listeners are forwarded as NetworkEvent.

Fields
firewall_channel: Sender<FirewallRequest<TRq>>requests_channel: EventChannel<ReceiveRequest<Rq, Rs>>events_channel: Option<EventChannel<NetworkEvent>>ident: Option<(AuthenticKeypair<X25519Spec>, PeerId)>behaviour_config: NetBehaviourConfig<TRq>connections_limit: Option<ConnectionLimits>
Implementations

impl<Rq, Rs, TRq> StrongholdP2pBuilder<Rq, Rs, TRq> where
Rq: RqRsMessage + Borrow<TRq>,
Rs: RqRsMessage,
TRq: Clone + Send + 'static,

pub fn new(
firewall_channel: Sender<FirewallRequest<TRq>>,
requests_channel: EventChannel<ReceiveRequest<Rq, Rs>>,
events_channel: Option<EventChannel<NetworkEvent>>
) -> Self

Parameters:

firewall_channel: Channel for FirewallRequest if there are no fixed rules in the firewall or Rule::Ask was set.

requests_channel: Channel for forwarding inbound requests from remote peers

events_channel: Optional channel for forwarding all events in the swarm.

pub fn with_keys(self, keys: InitKeypair) -> Self

Set the keypair that is used for authenticating the communication on the transport layer. The local PeerId is derived from the keypair.

pub fn with_connections_limit(self, limit: ConnectionLimits) -> Self

Set the limit for simultaneous connections. By default no connection limits apply.

pub fn with_request_timeout(self, t: Duration) -> Self

Set a timeout for receiving a response after a request was sent.

This applies for inbound and outbound requests.

pub fn with_connection_timeout(self, t: Duration) -> Self

Set the timeout for a idle connection to a remote peer.

pub fn with_firewall_config(self, config: FirewallConfiguration<TRq>) -> Self

Set the firewall configuration. The peer-specific rules overwrite the default rules for that peer.

Per default, no rules are set and a FirewallRequest::PeerSpecificRule request is sent through the firewall_channel when a peer connect or an inbound/ outbound request is sent.

pub async fn build(self) -> Result<StrongholdP2p<Rq, Rs, TRq>, Error>

Self::build_with_transport with a [Transport] based on TCP/IP that supports dns resolution and websockets. It uses [tokio::spawn] as executor, hence this method has to be called in the context of a tokio.rs runtime.

pub async fn build_with_transport<Tp, E>(
self,
transport: Tp,
executor: E
) -> StrongholdP2p<Rq, Rs, TRq> where
Tp: Transport + Sized + Clone + Send + Sync + 'static,
Tp::Output: AsyncRead + AsyncWrite + Unpin + Send + 'static,
Tp::Dial: Send + 'static,
Tp::Listener: Send + 'static,
Tp::ListenerUpgrade: Send + 'static,
Tp::Error: Send + Sync,
E: Executor + Send + 'static + Clone,

Create a new StrongholdP2p instance with an underlying [Swarm][libp2p::Swarm] that uses the provided transport.

The transport is upgraded with:

[Relay protocol][https://docs.libp2p.io/concepts/circuit-relay/] (requires feature = “relay”)

Authentication and encryption with the Noise-Protocol, using the XX-handshake

Yamux substream multiplexing

The method spawns a new task with the provided executor, that handles all interaction with the Swarm. The task runs until StrongholdP2p is dropped, StrongholdP2p provides an interface to perform operations on the swarm-task. Additionally, the executor is used to configure the [SwarmBuilder::executor][libp2p::swarm::SwarmBuilder::executor].