Module p2p::interface::types

Structs

Listener

Active Listener of the local peer.

ReceiveRequest

Inbound Request from a remote peer. It is expected that a response will be returned through the response_rx channel, otherwise an OutboundFailure will occur at the remote peer.

RequestId

Unique Id for each request. Note: This ID is only local and does not match the request’s ID at the remote peer.

Enums

InboundFailure

Possible failures occurring in the context of receiving an inbound request and sending a response.

NetworkEvent

Events happening in the Network. Includes events about connection and listener status as well as potential failures when receiving request-response messages.

OutboundFailure

Possible failures occurring in the context of sending an outbound request and receiving the response.

RequestDirection

Direction of a request.

Traits

RqRsMessage

Trait for the generic request and response messages.

Type Definitions

SwarmEv