Struct p2p::Context

pub struct Context { /* fields omitted */ }

Network context.

Methods

impl Context

pub fn new(
local_sync_node: LocalSyncNodeRef,
pool_handle: CpuPool,
remote: Remote,
config: Config
) -> Result<Self, Box<Error>>

Creates new context with reference to local sync node, thread pool and event loop.

pub fn spawn<F>(&self, f: F) where
F: Future + Send + 'static,
F::Item: Send + 'static,
F::Error: Send + 'static,

Spawns a future using thread pool and schedules execution of it with event loop handle.

pub fn execute_after<F>(&self, duration: Duration, f: F) where
F: FnOnce() + 'static + Send,

Schedules execution of function in future. Use wisely, it keeps used objects in memory until after it is resolved.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn node_table_entries(&self) -> Vec<Node>

Returns addresses of recently active nodes. Sorted and limited to 1000.

pub fn update_node_table(&self, nodes: Vec<AddressEntry>)

Updates node table.

pub fn penalize_node(&self, addr: &SocketAddr)

Penalize node.

pub fn add_node(&self, addr: SocketAddr) -> Result<(), NodeTableError>

Adds node to table.

pub fn remove_node(&self, addr: SocketAddr) -> Result<(), NodeTableError>

Removes node from table.

pub fn autoconnect(context: Arc<Context>, handle: &Handle)

Every 10 seconds check if we have reached maximum number of outbound connections. If not, connect to best peers.

pub fn connect<T>(context: Arc<Context>, socket: SocketAddr) where
T: SessionFactory,

Connect to socket using given context.

pub fn connect_normal(context: Arc<Context>, socket: SocketAddr)

Important traits for Box<R>
impl<R> Read for Box<R> where
R: Read + ?Sized, impl<W> Write for Box<W> where
W: Write + ?Sized, impl<I> Iterator for Box<I> where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn accept_connection_future(
context: Arc<Context>,
stream: TcpStream,
socket: SocketAddr,
handle: &Handle,
config: NetConfig
) -> Box<Future<Item = (), Error = ()> + Send>

pub fn accept_connection(
context: Arc<Context>,
stream: TcpStream,
socket: SocketAddr,
config: NetConfig
)

pub fn listen(
context: Arc<Context>,
handle: &Handle,
config: NetConfig
) -> Result<Box<Future<Item = (), Error = ()> + Send>, Error>

Starts tcp server and listens for incomming connections.

pub fn on_message(
context: Arc<Context>,
channel: Arc<Channel>
) -> IoFuture<MessageResult<()>>

Called on incomming mesage.

pub fn send_to_peer<T>(
context: Arc<Context>,
peer: PeerId,
payload: &T,
serialization_flags: u32
) -> IoFuture<()> where
T: Payload,

Send message to a channel with given peer id.

pub fn send_message_to_peer<T>(
context: Arc<Context>,
peer: PeerId,
message: T
) -> IoFuture<()> where
T: AsRef<[u8]> + Send + 'static,

pub fn send<T>(
_context: Arc<Context>,
channel: Arc<Channel>,
message: T
) -> IoFuture<()> where
T: AsRef<[u8]> + Send + 'static,

Send message using given channel.

pub fn close_channel(&self, id: PeerId)

Close channel with given peer info.

pub fn close_channel_with_error(&self, id: PeerId, error: &Error)

Close channel with given peer info.

pub fn create_sync_session(
&self,
start_height: i32,
services: Services,
outbound_connection: OutboundSyncConnectionRef
) -> InboundSyncConnectionRef

pub fn connections(&self) -> &Connections

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn nodes(&self) -> Vec<Node>