Trait p2p::Executor

pub trait Executor {
    fn exec(
        &self, 
        future: Pin<Box<dyn Future<Output = ()> + Send + 'static, Global>>
    );
}

Implemented on objects that can run a Future in the background.

Note: While it may be tempting to implement this trait on types such as [futures::stream::FuturesUnordered], please note that passing an Executor is optional, and that FuturesUnordered (or a similar struct) will automatically be used as fallback by libp2p. The Executor trait should therefore only be about running Futures in the background.

Required methods

fn exec(
&self,
future: Pin<Box<dyn Future<Output = ()> + Send + 'static, Global>>
)

Run the given future in the background until it ends.