Crate p2p

Modules

behaviour

firewall

interface

libp2p_reexport

Macros

unwrap_or_return

Structs

AuthenticKeypair
A DH keypair that is authentic w.r.t. a [identity::PublicKey].

ConnectionLimits
The configurable connection limits.

EstablishedConnections
Information about the connection with a remote peer as maintained in the ConnectionManager.

EventChannel
Wrapper of a [mpsc::channel][futures::channel::mpsc::channel] for sending events. The EventChannel cen be configure its behaviour e capacity of the channel is reached.

Listener
Active Listener of the local peer.

MessageProtocol
Protocol Name. A Request-Response messages will only be successful if both peers support the MessageProtocol.

Multiaddr
Representation of a Multiaddr.

NoiseKeypair
DH keypair.

PeerId
Identifier of a peer of the network.

ReceiveRequest
Inbound Request from a remote peer. It is expected that a response will be returned through the response_rx channel, otherwise an OutboundFailure will occur at the remote peer.

RequestId
Unique Id for each request. Note: This ID is only local and does not match the request’s ID at the remote peer.

StrongholdP2p
Interface for the stronghold-p2p library to create a swarm, handle events and perform operations.

StrongholdP2pBuilder
Builder for new StrongholdP2p.

Enums

ChannelSinkConfig
Configure how the network should behave in the case that the capacity of the EventChannel is reached, i.g. if the channel is full. This is relevant in cases when the frequency of messages is larger than the frequency in which the [mpsc::Receiver][futures::channel::mpsc::Receiver] side reads from the stream.

ConnectionErr
Error on establishing a connection.

DialErr
Error on dialing a peer and establishing a connection.

DialError
The possible failures of [ExpandedSwarm::dial].

InboundFailure
Possible failures occurring in the context of receiving an inbound request and sending a response.

InitKeypair
Use existing keypair for authentication on the transport layer.

Keypair
Identity keypair of a node.

ListenErr
Error on listening on an address.

ListenRelayErr
Error on listening on a relayed address.

NetworkEvent
Events happening in the Network. Includes events about connection and listener status as well as potential failures when receiving request-response messages.

OutboundFailure
Possible failures occurring in the context of sending an outbound request and receiving the response.

RequestDirection
Direction of a request.

TransportErr
Error on the [Transport][libp2p::Transport].

Traits

Executor
Implemented on objects that can run a Future in the background.

RqRsMessage
Trait for the generic request and response messages.

Functions

assemble_relayed_addr
Assemble a relayed address for the target following the syntax <relay-addr>/p2p/<relay-id>/p2p-circuit/p2p/<target-id>. The address can be used to reach the target peer if they are listening on that relay.