Struct p2p::Context

pub struct Context { /* fields omitted */ }

Network context.

Methods

impl Context

fn new(local_sync_node: LocalSyncNodeRef,
pool_handle: CpuPool,
remote: Remote,
config: Config)
-> Result<Self, Box<Error>>

Creates new context with reference to local sync node, thread pool and event loop.

fn spawn<F>(&self, f: F) where F: Future + Send + 'static,
F::Item: Send + 'static,
F::Error: Send + 'static

Spawns a future using thread pool and schedules execution of it with event loop handle.

fn execute_after<F>(&self, duration: Duration, f: F) where F: FnOnce() + 'static + Send

Schedules execution of function in future. Use wisely, it keeps used objects in memory until after it is resolved.

fn node_table_entries(&self) -> Vec<Node>

Returns addresses of recently active nodes. Sorted and limited to 1000.

fn update_node_table(&self, nodes: Vec<Node>)

Updates node table.

fn add_node(&self, addr: SocketAddr) -> Result<(), NodeTableError>

Adds node to table.

fn remove_node(&self, addr: SocketAddr) -> Result<(), NodeTableError>

Removes node from table.

fn autoconnect(context: Arc<Context>, handle: &Handle)

Every 10 seconds check if we have reached maximum number of outbound connections. If not, connect to best peers.

fn connect<T>(context: Arc<Context>, socket: SocketAddr) where T: SessionFactory

Connect to socket using given context.

fn connect_normal(context: Arc<Context>, socket: SocketAddr)

fn accept_connection_future(context: Arc<Context>,
stream: TcpStream,
socket: SocketAddr,
handle: &Handle,
config: NetConfig)
-> BoxFuture<(), ()>

fn accept_connection(context: Arc<Context>,
stream: TcpStream,
socket: SocketAddr,
config: NetConfig)

fn listen(context: Arc<Context>,
handle: &Handle,
config: NetConfig)
-> Result<BoxFuture<(), ()>, Error>

Starts tcp server and listens for incomming connections.

fn on_message(context: Arc<Context>,
channel: Arc<Channel>)
-> IoFuture<MessageResult<()>>

Called on incomming mesage.

fn send_to_peer<T>(context: Arc<Context>,
peer: PeerId,
payload: &T)
-> IoFuture<()> where T: Payload

Send message to a channel with given peer id.

fn send_message_to_peer<T>(context: Arc<Context>,
peer: PeerId,
message: T)
-> IoFuture<()> where T: AsRef<[u8]> + Send + 'static

fn send<T>(_context: Arc<Context>,
channel: Arc<Channel>,
message: T)
-> IoFuture<()> where T: AsRef<[u8]> + Send + 'static

Send message using given channel.

fn close_channel(&self, id: PeerId)

Close channel with given peer info.

fn close_channel_with_error(&self, id: PeerId, error: &Error)

Close channel with given peer info.

fn create_sync_session(&self,
start_height: i32,
outbound_connection: OutboundSyncConnectionRef)
-> InboundSyncConnectionRef

fn connections(&self) -> &Connections

fn nodes(&self) -> Vec<Node>